+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release Notes
 
 
+## Version 0.55.0
+**Release Date**: 4th October 2023
+### Features
+1. Extend PCA Query to lookup Images in all Tenancy Compartments.
+2. Remove unsupported DNS Zone icon.
+3. Prep for C3 Release.
+### Bug Fixes
+1. Resolve issue where availability domain was not passed as integer to the subnet.jinja2
+
+
 ## Version 0.54.5
 **Release Date**: 22nd September 2023
 ### Features